Fix tokenizer counting
The tokenizer included in Betalgo uses an older tokenizer which is incorrect for gpt-3.5 and higher. To fix this I'm now using Tiktoken which is currently the fastest port of the official Python version of the library.

Also OpenAI adds additional tokens which must be accounted for when using the Chat Completions API. This is now accounted for.